Added tickbox to make people check pending edits.
I think I have done this right. It makes no difference to the admin or anything like that, just something needed to try and nudge people to pending edits rather than lots of people submitting the same edits which clogs up the pending edits. 

Assistance needed to ass the pending edits link to the help text or into the label text. I could just add the normal pending edits link but that would be too long, ideally want the pending edits with the 'company NOC Code' filter on it.