Required proof
Not sure if this is right, but it feels right.

Hopefully adds a 'required' tick box to try and nudge people to provide proof of edits. A few too many edits are being done then later reverted as a number of users are making edits based off 'it hasn't been seen in so long so this must have happened to it' or random other excuses. If proof is requested, hopefully it will reduce some of the tooing and froing that we saw recently with a few buses.