TFLO TMP machines default withdraw
As TMP machines now link to the TFLO operator code and they are locked, it means they are forever visible as no one can withdraw them. 

This commit withdraws them by default as no one needs to edit them (bar withdraw) and no one needs to be able to see them on a list.