Updates operator_vehicles view to make it more user friendly
Currently when clicking from 'routes' to 'vehicles' the page jumps higher because there is no 'a [mode] operator in [region]' so I have added that in so the page doesn't jump when going between vehicle and routes view.

Also added 'tickets' link to Vehicles page where tickets link exists. Currently this is a bit messy as you can only access the 'tickets' page if you are on the 'routes' page and can't access it from the vehicles page.